+++ Piafometre.vcxproj
@@ -146,6 +146,9 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="Jeu.h" />
+    <ClInclude Include="moteur.h" />
+    <ClInclude Include="MoteurMenu.h" />
     <ClInclude Include="ressources.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
